customer ride cancellation

+++ src/main/java/com/alpha/RideAxis/Repository/CustomerRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.alpha.RideAxis.Entites.Customer;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
-	Optional<Customer> findByMobileno(long mobileno);  
-	void deleteByMobileno(long mobileno);
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+    Optional<Customer> findByMobileno(long mobileno);  
+    void deleteByMobileno(long mobileno);
     boolean existsByMobileno(long mobileno);
 }
